requirements.txt
Django==6.0.2
djangorestframework==3.16.1
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
pillow==12.1.1
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0